# docs/swagger.yaml
openapi: 3.0.0
info:
  title: CarLedgr API
  description: |
    Comprehensive dealership management system API for vehicle repair and resale businesses.
    
    **Authentication:** Most endpoints require JWT token in Authorization header: `Bearer <token>`
    
    **Organization Isolation:** All data is automatically scoped to the authenticated user's organization
    
    **License Enforcement:** Some endpoints check subscription status and car limits
  version: 2.0.0
  contact:
    name: CarLedgr API Support
    url: https://carledgr.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3030/api
    description: Development server
  - url: https://demo-api.carledgr.com/api
    description: Demo server
  - url: https://api.carledgr.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: Get all organizations (Owner only)
      responses:
        '200':
          description: List of organizations

    post:
      tags: [Organizations]
      summary: Create organization (Owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganization'
      responses:
        '201':
          description: Organization created
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization by ID (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization details

    put:
      tags: [Organizations]
      summary: Update organization (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization updated

    delete:
      tags: [Organizations]
      summary: Delete organization (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization deleted

  # Stripe Integration
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateOrganization:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
...